Updated the insertAssertions and insertCustomAssertions commands to insert the corresponding assertions whenever multiple vtrace() calls are identified.
At the moment, the script looks at the code above the line where the command was called to identify the most recent vtrace() call and will only insert the assertions corresponding to that vtrace() from DIG's output.

Currently working on only allowing the user to call the commands on a line where a vtrace() call is present, because calling them on any line of code below that will insert the assertions generated by that vtrace() call, which might not hold a few lines below the vtrace() call.