Completed the Test Assertion function
I completed the Test Assertion function. It uses the subprocess.run(...) method inside CIVL and writes the output to a JSON file. The script then reads that JSON file and looks for the "Violation" keyword to determine whether an assertion is valid or not. I am using that keyword because it only appears when an assertion is not valid. I am going to do some extensive testing to refine and ensure the feature works as intended.